refactor(codegen): extrai calls/indirect.rs
Move 3 fns de chamadas indiretas (~282 LOC) para
expressions/calls/indirect.rs:
- lower_var_member_call: obj.method(args) generico (sem tipo class)
- lower_indirect_call: callee Ident apontando pra valor (param com func_addr)
- emit_function_handle_indirect_call: invoke via handle Function (.call/.bind)

calls/mod.rs: 1818 -> 1541 linhas. Suite verde.
